+++ packages/core/src/debug/ai/signal_graph.ts
@@ -9,7 +9,7 @@
 import {NullInjector} from '../../di/null_injector';
 import {getInjector} from '../../render3/util/discovery_utils';
 import {DebugSignalGraph, getSignalGraph} from '../../render3/util/signal_debug';
-import {ToolDefinition} from './tool_definitions';
+import {ToolDefinition} from '../../../primitives/devtools';
 
 // Omit `debuggableFn` and `id` from returned signal graph to AI agent.
 type AiSignalGraph = Omit<DebugSignalGraph, 'nodes'> & {
